Move getGlobalSharedCache to base class
