+++ include/tc/core/compilation_cache-inl.h
@@ -15,12 +15,96 @@
  */
 #pragma once
 
+#include <sys/stat.h>
 #include <algorithm>
+#include <fstream>
 #include <iostream>
+#include <string>
 
+#include <glog/logging.h>
+
+#include <version.h>
 #include "tc/core/utils/math.h"
 
 namespace tc {
+
+template <typename CC, typename CachedEntryType>
+void Cache<CC, CachedEntryType>::enableCache() {
+  getGlobalSharedCache() = std::make_shared<CC>();
+}
+
+template <typename CC, typename CachedEntryType>
+void Cache<CC, CachedEntryType>::disableCache() {
+  getGlobalSharedCache() = nullptr;
+}
+
+template <typename CC, typename CachedEntryType>
+std::shared_ptr<CC> Cache<CC, CachedEntryType>::getCache() {
+  if (not cacheEnabled()) {
+    throw std::runtime_error(
+        "EnableCache or LoadCacheFromProtobuf must be called before using the cache.");
+  }
+  return getGlobalSharedCache();
+}
+
+template <typename CC, typename CachedEntryType>
+std::shared_ptr<CC>& Cache<CC, CachedEntryType>::getGlobalSharedCache() {
+  static std::shared_ptr<CC> cache_ = nullptr;
+  return cache_;
+}
+
+template <typename CC, typename CachedEntryType>
+void Cache<CC, CachedEntryType>::dumpCacheToProtobuf(
+    const std::string& filename) {
+  std::fstream serialized(
+      filename, std::ios::binary | std::ios::trunc | std::ios::out);
+  if (!serialized) {
+    LOG(ERROR) << "Failed to open the output stream for dumping protobuf: "
+               << filename;
+  } else {
+    getCache()->toProtobuf().SerializePartialToOstream(&serialized);
+  }
+}
+
+template <typename CC, typename CachedEntryType>
+void Cache<CC, CachedEntryType>::loadCacheFromProtobuf(
+    const std::string& filename) {
+  typename CC::ProtobufType buf;
+  struct stat buffer = {0};
+  if (stat(filename.c_str(), &buffer) == 0) {
+    std::ifstream serialized(filename, std::ios::binary);
+    buf.ParseFromIstream(&serialized);
+  }
+  loadCacheFromProtobuf(buf);
+}
+
+template <typename CC, typename CachedEntryType>
+template <typename Protobuf>
+void Cache<CC, CachedEntryType>::loadCacheFromProtobuf(const Protobuf& buf) {
+  static_assert(
+      std::is_same<Protobuf, typename CC::ProtobufType>::value,
+      "LoadCacheFromProtobuf called with invalide protobuf type.");
+  CC::getGlobalSharedCache() = std::make_shared<CC>(buf);
+}
+
+template <typename CC, typename CachedEntryType>
+bool Cache<CC, CachedEntryType>::cacheEnabled() {
+  return getGlobalSharedCache() != nullptr;
+}
+
+template <typename CC, typename CachedEntryType>
+size_t Cache<CC, CachedEntryType>::size() const {
+  std::lock_guard<std::mutex> lock(mtx_);
+  return static_cast<const CC*>(this)->entries_.size();
+}
+
+template <typename CC, typename CachedEntryType>
+void Cache<CC, CachedEntryType>::clear() {
+  std::lock_guard<std::mutex> lock(mtx_);
+  numberAttemptedRetrievals = numberSuccessfulRetrievals = numberCacheAttemps =
+      0;
+  static_cast<CC*>(this)->entries_.clear();
+}
 
 namespace detail {
 template <typename CachedEntryType, typename TensorType>
@@ -230,13 +314,6 @@ OptionsCacheEntryProto OptionsCachedEntry<MappingOptionsType>::toProtobuf()
         return buf;
       });
   return buf;
-}
-
-template <typename MappingOptionsType>
-std::shared_ptr<OptionsCache<MappingOptionsType>>&
-OptionsCache<MappingOptionsType>::getGlobalSharedCache() {
-  static std::shared_ptr<OptionsCache> optionsCache_;
-  return optionsCache_;
 }
 
 template <typename MappingOptionsType>
